move: pther specific components to their respective dirs
