+++ apps/web/app/(app)/environments/[environmentId]/surveys/[surveyId]/(analysis)/summary/StatusDropdown.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import LoadingSpinner from "@/components/shared/LoadingSpinner";
-import SurveyStatusDropdown from "@/components/shared/SurveyStatusDropdown";
+import SurveyStatusDropdown from "@/app/(app)/environments/[environmentId]/surveys/[surveyId]/SurveyStatusDropdown";
 import { useEnvironment } from "@/lib/environments/environments";
 import { TSurvey } from "@formbricks/types/v1/surveys";
 import { ErrorComponent } from "@formbricks/ui";
